feat: consolidate guess generation
