+++ src/metastable/paths/path_solvers.py
@@ -6,10 +6,9 @@ from typing import Tuple
 from scipy.integrate import solve_bvp
 from scipy.integrate._bvp import BVPResult
 
-from metastable.map.map import FixedPointMap, FixedPointType, PathType
+from metastable.map.map import FixedPointMap, FixedPointType
 from metastable.eom import EOM
 from metastable.generate_boundary_conditions import generate_boundary_condition_func
-from metastable.generate_guess import generate_guess_from_sol
 
 from metastable.paths.data_structures import IndexPair
 from metastable.paths.parameter_utils import extract_params, prepare_saddle_and_focus_points
